Add Dedup function
Add a function to dedup an set of RRs. For this to work nicely I needed
something to format an RR with a Sprintf like function. So add that too.
For the Sprintf to work a rdataString() for each RR would be nice: add
that too. The %r modifier in Sprintf allows you to access this.

Don't really want to add RdataString() as an exported function to each
RR, as this add a whole load of extra methods. So just a Sprintf
function.

Also add Dedup function that uses all this.