Enables checking of tranports on a per-Msg basis.
In order to not be a participant of DNS amplification attacks,
one needs to limit record sizes for UDP connections.
One solution would be to curry the function for the Handler,
but this would mean i need different muxes for different listeners.

To avoid this uglyness, i added w.Tcp() and w.Udp() to determine
wether the writer's transport is either or, thus being able to limit
response sizes easily and conditionally.