REUSE_PORT: define in Linux
For some reason syscall.SO_REUSEPORT is not defined for Linux/amd64,386.
Hack around this by taking the value 0x0200 and play with build tags.

Also add code that sets these values on the socket.

I've only locally tested this with an UDP server, don't know if there
are any regression for TCP (or improvements)