return the matched string to the callback
When a dns server is handling incoming client requests, it
matches the incoming Question to a string stored in a map
that maps configured strings to callbacks. It can be very
useful for those callbacks to be informed of the configured
string that matched instead of having to repeat that logic.

To do so, the callback from the dns package needs to be
updated to add the string and it needs to be done in a
backwards compatible way. Do this by adding another callback
type that will return a structure so that if the callback
needs to change in the future it's as simple as adding a
field to the structure. gpg failed to sign the data fatal:
failed to write commit object

Signed-off-by: Keith Holleman <holleman@skyportsystems.com>