Fix go-routine leak
When and higher level (grammar or syntax) error was encountered when
parsing records from strings the routine would return, but the openend
channel would be left open and no other error would be triggered; this
means the goroutine starting z.lexer would be left hanging.

The fix is a bit of a hack; as we signal the upper error back down via
s.e and the trigger an "error" in the lexer, this at least ensure the
z.lexer function return and cleans up the goroutine.

Fixes #586
Fixes https://github.com/coredns/coredns/issues/1233