Restructure pull-request to avoid breaking the API
This commit removes the interface{} 'meta' argument, which required
breaking the signature of TsigSecret, and instead passes the name
directly to the callback. If there is a need to use a non-string secret,
this can be looked up from within a closure using the name as a lookup
key.

This change still permits all the same use-cases as it's parent.