Server: check client udp buf size
If we are returning a large response we now just send it (for udp) even
if the clients advertised buffer will be too small to take it all.
Make the response struct keep track of this udp size and cap returns
that exceed this.

Note this is irrespective of the TC bit in the message; we truncate
regardless.

This fails a test; but the test is wrong.