DOH helper functions
IETF hackathon

I was trying to make DNS over HTTPs a first class citizen, i.e. create a
dns.Conn and set Net to https (basically just like DNS over TLS), but
this ran into a few snags:

* http/2 has it's own socket pipelining so I can't give it a net.Conn
  and make it run with it. This breaks an assumption in this library.
* The Go http client setup has a function called client.Do which sends
  a reply and waits for a reponse, ala Exchange() in this library. But
  this library uses conn.WriteMsg and conn.ReadMsg as two calls to
  implement that - both of these are also public. So I could fake one of
  these in the case of http/2, i.e. make WriteMsg a noop and do all the
  work in ReadMsg, but I still don't have access to the conn used in the
  http/2 - so that won't fly either.

In the end I ended up with some helper functions that parse from and to
the http/2 request world, but nothing more tangible can be done in this
library. In this regard it is similar to the gRPC support we have in
CoreDNS which is also outside of this library.