Add support for truncating a message to fit a given buffer size
If DNS messages are too large for a given (UDP) buffer size, the
truncated bit should be set to request clients to retry the request via
a different transport (TCP).

This change introduces a Truncate() function which can be used to handle
the modification of the message. As some basic clients (incorrectly)
don't handle the truncation bit, it aims to return as many records as
possible while still being performant (e.g. not traversing resource
record lists multiple times to find the maximum number of fitting
records).

It ensures to always keep TSIG and OPT records in the response according
to the spec (e.g. RFC 6891 section 7).