Fix all cases of error halding
Use "err" if we are in a if-clause, if not start with e and then e1, e2,
etc. If this is be sure to never run into the situation where the error
interface has been given a type and thus can never be compared with nil
again: https://golang.org/doc/faq#nil_error

Something more elegant might be needed in the long run. There is also a
vague plan to autogenerate this - problem there is the weirdness some
RRs for parsing (depending on previous values, etc. etc.) Not that would
solve this issue, but does stop people from having to edit this code
directly ever again.

Signed-off-by: Miek Gieben <miek@miek.nl>