When using TCP connections send only one packet
net.Buffers relies on writeBuffer to send all the pieces in one go. The
tls library does not support this so the length and content get split
into 2 packets. This patch removes net.Buffers and replaces with a new
single memory allocation to ensure the length+content get sent in one
packet. This fixes #1199