Add MDNS class helper functions to RR_Header and Question
Add two methods to be used with MDNS (RFC6762) messages for dealing
with extra data encoded in to the class field of queries and resource
record headers.

For queries the top bit of the class field is the unicast-response bit
indicating the querier can accept a unicast response to the queuery (in
addition or in lieu of a multicast response)

For resource records the top bit of the class field is for the
cache-flush bit that indicates the record is unique and should replace
any entries vs appending to the cache.

Fixes some of the discussed parts of #436 and #423