Fix go generate
This adds the needed fixed for 'go generate', allthough I couldn't debug
running 'go generate' itself, but had to run the 3 *_generate.go
programs in sequence.

Added the AMTRELAY type code to test if this actually generated
something.

Added a check in the generate programs if enough code is generate as not
to globber the existing files if they fail to work.

Move to the new packages API and updated go.sum as go generate was
complaining about that as well.

Run the *_generate.go's in the CI.

Signed-off-by: Miek Gieben <miek@miek.nl>